requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pycryptodome==3.19.0
eth-account==0.10.0
loguru==0.7.3
pytest==8.4.1
pytest-cov==6.2.1
web3==7.12.0